+++ spring-rabbit-test/src/test/java/org/springframework/amqp/rabbit/test/examples/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * JUnit test examples.
+ */
+package org.springframework.amqp.rabbit.test.examples;
